+++ package.json
@@ -20,6 +20,7 @@
     "core-js": "^3.6.5",
     "firebase": "^7.14.0",
     "dayjs": "^1.8.24",
+    "reflect-metadata": "^0.1.13",
     "tsyringe": "^4.1.0",
     "vue": "^2.6.11",
     "vue-router": "^3.1.6",
