+++ src/main.ts
@@ -1,3 +1,4 @@
+import "reflect-metadata";
 import Vue from "vue";
 import VueCompositionApi from "@vue/composition-api";
 import App from "./App.vue";
