+++ src/views/Home.vue
@@ -1,18 +1,18 @@
 <template>
   <div class="home">
     <img alt="Vue logo" src="../assets/logo.png" />
-    <HelloWorld msg="Welcome to Your Vue.js App" />
+    <TodaysTasks />
   </div>
 </template>
 
 <script>
 // @ is an alias to /src
-import HelloWorld from "@/components/HelloWorld.vue";
+import TodaysTasks from "@/components/TodaysTasks.vue";
 
 export default {
   name: "Home",
   components: {
-    HelloWorld,
+    TodaysTasks,
   },
 };
 </script>
